feature/searchgroups -> filtro para pesquisa de grupos feito.
